requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
cloudinary==1.36.0
gunicorn==21.2.0
Werkzeug==3.0.1